+++ asdf/yamlutil.py
@@ -1,5 +1,6 @@
 import warnings
 from collections import OrderedDict
+from types import GeneratorType
 
 import numpy as np
 
@@ -212,21 +213,63 @@ AsdfLoader.add_constructor(None, AsdfLoader.construct_undefined)
 AsdfLoader.add_constructor(YAML_TAG_PREFIX + "omap", AsdfLoader.construct_yaml_omap)
 
 
-def custom_tree_to_tagged_tree(tree, ctx):
+def custom_tree_to_tagged_tree(tree, ctx, _serialization_context=None):
     """
     Convert a tree, possibly containing custom data types that aren't
     directly representable in YAML, to a tree of basic data types,
     annotated with tags.
     """
-    def walker(node):
-        tag = ctx.type_index.from_custom_type(type(node), ctx.version_string)
-        if tag is not None:
-            return tag.to_tree_tagged(node, ctx)
-        return node
+    if _serialization_context is None:
+        _serialization_context = ctx._create_serialization_context()
+
+    extension_manager = _serialization_context.extension_manager
+
+    def _convert_obj(obj):
+        converter = extension_manager.get_converter_for_type(type(obj))
+        tag = converter.select_tag(obj, _serialization_context)
+        node = converter.to_yaml_tree(obj, tag, _serialization_context)
+
+        if isinstance(node, GeneratorType):
+            generator = node
+            node = next(generator)
+        else:
+            generator = None
+
+        if isinstance(node, dict):
+            tagged_node = tagged.TaggedDict(node, tag)
+        elif isinstance(node, list):
+            tagged_node = tagged.TaggedList(node, tag)
+        elif isinstance(node, str):
+            tagged_node = tagged.TaggedString(node)
+            tagged_node._tag = tag
+        else:
+            raise TypeError(
+                "Converter returned illegal node type: {}".format(util.get_class_name(node))
+            )
+
+        _serialization_context._mark_extension_used(converter.extension)
+
+        yield tagged_node
+        if generator is not None:
+            yield from generator
+
+    def _walker(obj):
+        if extension_manager.handles_type(type(obj)):
+            return _convert_obj(obj)
+        else:
+            tag = ctx.type_index.from_custom_type(
+                type(obj),
+                ctx.version_string,
+                _serialization_context=_serialization_context
+            )
+
+            if tag is not None:
+                return tag.to_tree_tagged(obj, ctx)
+            return obj
 
     return treeutil.walk_and_modify(
         tree,
-        walker,
+        _walker,
         ignore_implicit_conversion=ctx._ignore_implicit_conversion,
         # Walk the tree in preorder, so that extensions can return
         # container nodes with unserialized children.
@@ -235,12 +278,17 @@ def custom_tree_to_tagged_tree(tree, ctx):
     )
 
 
-def tagged_tree_to_custom_tree(tree, ctx, force_raw_types=False):
+def tagged_tree_to_custom_tree(tree, ctx, force_raw_types=False, _serialization_context=None):
     """
     Convert a tree containing only basic data types, annotated with
     tags, to a tree containing custom data types.
     """
-    def walker(node):
+    if _serialization_context is None:
+        _serialization_context = ctx._create_serialization_context()
+
+    extension_manager = _serialization_context.extension_manager
+
+    def _walker(node):
         if force_raw_types:
             return node
 
@@ -248,7 +296,13 @@ def tagged_tree_to_custom_tree(tree, ctx, force_raw_types=False):
         if tag is None:
             return node
 
-        tag_type = ctx.type_index.from_yaml_tag(ctx, tag)
+        if extension_manager.handles_tag(tag):
+            converter = extension_manager.get_converter_for_tag(tag)
+            obj = converter.from_yaml_tree(node.data, tag, _serialization_context)
+            _serialization_context._mark_extension_used(converter.extension)
+            return obj
+
+        tag_type = ctx.type_index.from_yaml_tag(ctx, tag, _serialization_context=_serialization_context)
         # This means the tag did not correspond to any type in our type index.
         if tag_type is None:
             if not ctx._ignore_unrecognized_tag:
@@ -283,7 +337,7 @@ def tagged_tree_to_custom_tree(tree, ctx, force_raw_types=False):
 
     return treeutil.walk_and_modify(
         tree,
-        walker,
+        _walker,
         ignore_implicit_conversion=ctx._ignore_implicit_conversion,
         # Walk the tree in postorder, so that extensions receive
         # container nodes with children already deserialized.
@@ -304,7 +358,7 @@ def load_tree(stream):
     return yaml.load(stream, Loader=AsdfLoader)
 
 
-def dump_tree(tree, fd, ctx):
+def dump_tree(tree, fd, ctx, tree_finalizer=None, _serialization_context=None):
     """
     Dump a tree of objects, possibly containing custom types, to YAML.
 
@@ -318,11 +372,17 @@ def dump_tree(tree, fd, ctx):
 
     ctx : Context
         The writing context.
-    """
-    class AsdfDumperTmp(AsdfDumper):
-        pass
-    AsdfDumperTmp.ctx = ctx
 
+    tree_finalizer : callable, optional
+        Callback that receives the tagged tree before it is validated
+        and defaults are removed.  `asdf.AsdfFile` uses this to update
+        extension metadata on the tagged tree before it is fully
+        serialized to YAML.
+    """
+    # The _serialization_context parameter allows AsdfFile to track
+    # what extensions were used when converting the tree's custom
+    # types.  In 3.0, it will be passed as the `ctx` instead of the
+    # AsdfFile itself.
     tags = None
     tree_type = ctx.type_index.from_custom_type(type(tree))
     if tree_type is not None:
@@ -335,15 +395,16 @@ def dump_tree(tree, fd, ctx):
         yaml_tag = ':'.join(tag_parts[0:-1] + [last_part])
         tags = {'!': yaml_tag}
 
-    tree = custom_tree_to_tagged_tree(tree, ctx)
+    tree = custom_tree_to_tagged_tree(tree, ctx, _serialization_context=_serialization_context)
+    if tree_finalizer is not None:
+        tree_finalizer(tree)
     schema.validate(tree, ctx)
-    schema.remove_defaults(tree, ctx)
 
     yaml_version = tuple(
         int(x) for x in ctx.version_map['YAML_VERSION'].split('.'))
 
     yaml.dump_all(
-        [tree], stream=fd, Dumper=AsdfDumperTmp,
+        [tree], stream=fd, Dumper=AsdfDumper,
         explicit_start=True, explicit_end=True,
         version=yaml_version,
         allow_unicode=True, encoding='utf-8',